refactor(storage): extract theme handling into a composable and reuse it across components
